% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erviss_variants.R
\name{show_variants_for_a_given_period}
\alias{show_variants_for_a_given_period}
\title{Show variants for a given period}
\usage{
show_variants_for_a_given_period(
  csv_variants_file,
  date_min,
  date_max,
  variant_to_study = "",
  studysites_variants = "",
  minimal_value = 0,
  indicator_to_study = "proportion",
  date_breaks,
  date_format = "\%b \%Y"
)
}
\arguments{
\item{csv_variants_file}{The path to the CSV file containing the variants data}

\item{date_min}{The minimum date to show}

\item{date_max}{The maximum date to show}

\item{variant_to_study}{The variants to study}

\item{studysites_variants}{The studysites to study}

\item{minimal_value}{The minimal value to show}

\item{indicator_to_study}{The indicator to study}

\item{date_breaks}{The breaks for the x-axis}

\item{date_format}{The format for the x-axis}
}
\value{
A plot of the variants
}
\description{
Show variants for a given period
}
\examples{
show_variants_for_a_given_period(
  csv_variants_file = "https://raw.githubusercontent.com/EU-ECDC/Respiratory_viruses_weekly_data/refs/heads/main/data/snapshots/2025-11-21_variants.csv",
  date_min = as.Date("2024-10-01"),
  date_max = as.Date("2025-09-30"),
  variant_to_study = c("XFG", "LP.8.1", "BA.2.86"),
  minimal_value = 10,
  studysites_variants =  c("Belgium", "Denmark", "Italy", "Norway", "Portugal", "Spain", "Sweden"),
  date_breaks = "1 month", 
  date_labels = "\%b \%Y"
)
}
