% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erviss_variants.R
\name{clean_erviss_variants_for_a_given_period}
\alias{clean_erviss_variants_for_a_given_period}
\title{Clean the variants data for a given period}
\usage{
clean_erviss_variants_for_a_given_period(
  csv_variants_file,
  date_min,
  date_max,
  variant_to_study = "",
  studysites_variants = "",
  minimal_value = 0,
  indicator_to_study = "proportion"
)
}
\arguments{
\item{csv_variants_file}{The path to the CSV file containing the variants data}

\item{date_min}{The minimum date to show}

\item{date_max}{The maximum date to show}

\item{variant_to_study}{The variants to study}

\item{studysites_variants}{The studysites to study}

\item{minimal_value}{The minimal value to show}

\item{indicator_to_study}{The indicator to study}
}
\value{
A data frame containing the variants data
#' @export
}
\description{
Clean the variants data for a given period
}
