% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erviss_positivity.R
\name{show_positivity_for_a_given_period}
\alias{show_positivity_for_a_given_period}
\title{Show variants for a given period}
\usage{
show_positivity_for_a_given_period(
  csv_variants_file,
  date_min,
  date_max,
  pathogen_to_study = "",
  studysites_variants = ""
)
}
\arguments{
\item{csv_variants_file}{The path to the CSV file containing the variants data}

\item{date_min}{The minimum date to show}

\item{date_max}{The maximum date to show}

\item{studysites_variants}{The studysites to study}

\item{variant_to_study}{The variants to study}
}
\value{
A plot of the variants
}
\description{
Show variants for a given period
}
